refactor(#58) :: home의 Calendar 폴더 구조 리펙토링
