+++ src/features/Home/Calendar/index.ts
@@ -1,4 +1,2 @@
-export { default as CalendarPlus } from './CalendarPlus';
-export { default as CalendarSearch } from './CalendarSearch';
 export { default as CalendarModal } from './CalendarModal';
 export { default as CalendarToggle } from './CalendarToggle';
