+++ src/features/Home/ui/index.ts
@@ -2,3 +2,9 @@ export { default as BtnToDoNormal } from './BtnToDoNormal';
 export { default as BtnToDoChecked } from './BtnToDoChecked';
 export { default as BtnCategoryPlus } from './BtnCategoryPlus';
 export { default as BtnCategoryItem } from './BtnCategoryItem';
+export { default as CalendarPlusIcon } from './CalendarPlusIcon';
+export { default as CalendarSearchIcon } from './CalendarSearchIcon';
+export { default as DropdownCheckIcon } from './DropdownCheckIcon';
+export { default as DropdownArrowIcon } from './DropdownArrow';
+export { default as CalendarScheduleIcon } from './CalendarScheduleIcon';
+export { default as CalendarTodoIcon } from './CalendarTodoIcon';
