--- src/pages/Home/ui/Calendar/ui/CalendarToggle/ui/index.ts
@@ -1,2 +1,0 @@
-export { default as CalendarSchedule } from './CalendarSchedule';
-export { default as CalendarTodo } from './CalendarTodo';
