--- src/pages/Home/ui/Calendar/ui/Dropdown/ui/index.ts
@@ -1,3 +1,0 @@
-export { default as DropdownCheck } from './DropdownCheck';
-export { default as DropdownArrow } from './DropdownArrow';
-export { default as DropdownItem } from './DropdownItem';
